todo-app/requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
python-dotenv
flask-login
werkzeug